Adding a new space_char argument for Trainer.
By default Trainer uses ' ' as `space_char`, that means
that in the default implementation `▁` is actually treated as a regular
char, which has a specific SCRIPT tag, which ruins the vocab.

We `need` to have a way to control `space_char` from the python side (in
case we don't use Metaspace). We can always use the standard
`replacement` as a default though.

The parity check actually currently fails on `big.txt` but the number of
tokens is in the same ballpark `153k vs 152k`. In order to do the check
with OK speed, we need to compile in `--release` mode which is not the
default. In the current commit I propose to simply disable `--debug`
builds has they are probably not useful to anyone.