Adding a new PreTokenizer for UnicodeScripts.
The core idea is to detect the "Script" of every char in a String
and split whenever there is a change of Script.

For instance

"どこで生れ。Yes" -> "どこで生れ", "。", "Yes"

The first chars are all considered `Han`, "。" is `Common` and the rest are
`Latin`.
This PreTokenizer will be helpful to disentangle pre-processing done in
`spm` and be able to reuse it better when needed.